third_party/blink/web_tests/dom/comment/remove.html

<!DOCTYPE html>
<title>This tests the DOM 4 remove method on a Comment</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="test"></div>
<script>
setup({single_test: true});
var testDiv = document.getElementById('test');
var comment = document.createComment('Comment');
testDiv.appendChild(comment);
assert_equals(testDiv.childNodes.length, 1);
comment.remove();
assert_equals(testDiv.childNodes.length, 0);
comment.remove();
assert_equals(testDiv.childNodes.length, 0);
done();
</script>
